requirements.txt
requests
fastapi
uvicorn
numpy
pillow
pydantic
Flask